# openapi/hilt-review-openapi.yaml
openapi: 3.1.0
info:
  title: HILT-Review API
  version: 2.0.0
  description: >
    Human-in-the-loop review gatekeeper for async headless agent pipelines.
    Agents propose actions as Review Tasks. Humans edit/approve/deny.
    HILT-Review emits decisions to downstream executors via per-source webhooks
    and/or a pull API. Delivery is at-least-once; consumers must dedupe using
    event IDs and idempotency keys.


    ## Authentication

    - **Machine clients** (agents, executors): Use API key via `X-API-Key` header
    - **Human reviewers**: Authenticate via Google OAuth 2.0 (session-based)


    ## Rate Limiting

    All endpoints are rate-limited. Limits are per API key or per session.
    When rate limited, the API returns `429 Too Many Requests`.

    | Endpoint Category | Limit |
    |-------------------|-------|
    | Task Creation | 100 requests/minute |
    | Task Reads | 1000 requests/minute |
    | Decision Events | 500 requests/minute |
    | Source Management | 50 requests/minute |

    Rate limit headers are included in all responses:
    - `X-RateLimit-Limit`: Maximum requests allowed in window
    - `X-RateLimit-Remaining`: Requests remaining in current window
    - `X-RateLimit-Reset`: Unix timestamp when window resets
servers:
  - url: https://api.hilt-review.example.com
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  headers:
...
    RateLimitLimit:
      description: Maximum number of requests allowed in the current window.
      schema:
        type: integer
        example: 100
    RateLimitRemaining:
      description: Number of requests remaining in the current window.
      schema:
        type: integer
        example: 95
    RateLimitReset:
      description: Unix timestamp (seconds) when the rate limit window resets.
      schema:
        type: integer
        example: 1699900060
  responses:
    BadRequest:
      description: Invalid request - malformed JSON, missing required fields, or validation errors.
      headers:
        X-RateLimit-Limit:
          $ref: "#/components/headers/RateLimitLimit"
        X-RateLimit-Remaining:
          $ref: "#/components/headers/RateLimitRemaining"
        X-RateLimit-Reset:
          $ref: "#/components/headers/RateLimitReset"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: "VALIDATION_ERROR"
              message: "Request validation failed"
              details:
                fields:
                  - field: "blocks[0].id"
                    message: "Block ID is required"
    Unauthorized:
      description: Missing or invalid authentication credentials.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: "UNAUTHORIZED"
              message: "Invalid or missing API key"
    Forbidden:
      description: Authenticated but not authorized for this operation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: "FORBIDDEN"
              message: "API key does not have access to this source"
...
    Conflict:
      description: Request conflicts with current state (idempotency mismatch, invalid state transition).
      headers:
        X-RateLimit-Limit:
          $ref: "#/components/headers/RateLimitLimit"
        X-RateLimit-Remaining:
          $ref: "#/components/headers/RateLimitRemaining"
        X-RateLimit-Reset:
          $ref: "#/components/headers/RateLimitReset"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: "IDEMPOTENCY_MISMATCH"
              message: "Idempotency key was used with different request body"
...
    TooManyRequests:
      description: Rate limit exceeded. Retry after the window resets.
      headers:
        X-RateLimit-Limit:
          $ref: "#/components/headers/RateLimitLimit"
        X-RateLimit-Remaining:
          schema:
            type: integer
            example: 0
        X-RateLimit-Reset:
          $ref: "#/components/headers/RateLimitReset"
        Retry-After:
          description: Seconds until the rate limit resets.
          schema:
            type: integer
            example: 30
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: "RATE_LIMITED"
              message: "Too many requests. Retry after 30 seconds."
              details:
                retry_after_seconds: 30
    InternalServerError:
      description: Unexpected server error. Safe to retry with exponential backoff.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: "INTERNAL_ERROR"
              message: "An unexpected error occurred"
              request_id: "req_abc123xyz"
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code.
              enum:
                - VALIDATION_ERROR
                - UNAUTHORIZED
                - FORBIDDEN
                - NOT_FOUND
                - IDEMPOTENCY_MISMATCH
                - PRECONDITION_FAILED
                - INVALID_STATE_TRANSITION
                - RATE_LIMITED
                - INTERNAL_ERROR
            message:
              type: string
              description: Human-readable error message.
            details:
              type: object
              description: Additional error context (varies by error type).
            request_id:
              type: string
              description: Request ID for debugging (included on server errors).

    # -------- Common Types --------
    UUID:
      type: string
      format: uuid
    Timestamp:
      type: string
      format: date-time
    Priority:
      type: string
      enum: [LOW, NORMAL, HIGH]
    RiskLevel:
      type: string
      enum: [low, medium, high, critical]
      description: >
        Risk level indicating consequence severity.
        - low: Easily reversible, private scope (drafts, notes, archive)
        - medium: Limited audience, some friction to undo (reply to known contact)
        - high: Broad audience, hard to undo (public post, external email)
        - critical: Irreversible, financial/legal (transactions, deletions)
    ServiceIdentifier:
      type: object
      description: Identifies the service/platform for visual display in UI.
      required: [id, name]
      properties:
        id:
          type: string
          description: Machine identifier (e.g., "gmail", "notion", "twitter", "slack").
        name:
          type: string
          description: Human-readable name (e.g., "Gmail", "Notion", "Twitter").
        icon:
          type: string
          description: Icon key or emoji for UI display (e.g., "✉️", "📓", "🐦").
    ActionIdentifier:
      type: object
      description: Identifies the action type for visual display in UI.
      required: [type, verb]
      properties:
        type:
          type: string
          enum: [send, reply, post, create, update, delete, archive, schedule, transfer, notify]
          description: Machine-readable action type.
        verb:
          type: string
          description: Human-readable verb for display (e.g., "Reply", "Post publicly", "Transfer $450").
        icon:
          type: string
          description: Icon key or emoji for UI display (e.g., "📤", "📢", "💰").
    ReviewTaskStatus:
      type: string
      enum: [PENDING, APPROVED, DENIED, DISPATCHED, ARCHIVED]
    DecisionType:
      type: string
      enum: [APPROVE, DENY]
    BlockType:
      type: string
      enum: [markdown, plaintext, json]
    ArtifactBlock:
      type: object
      required: [id, type, content, editable]
      properties:
        id:
          type: string
          description: Stable machine identifier for referencing/editing/diffing a block (e.g., "email_subject").
        label:
          type: string
          description: Human-readable label for UI display (e.g., "Email Subject"). Provided by agent.
        type:
          $ref: "#/components/schemas/BlockType"
        content:
          description: Block contents; string for markdown/plaintext, object for json.
          oneOf:
            - type: string
            - type: object
        editable:
          type: boolean
        render_hints:
          type: object
          properties:
            preview:
              type: boolean
            syntax_highlighting:
              type: boolean
    InteractionSchema:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [edit, confirm, choice, external]
        # Optional guidance layer; never removes editability.
        guidance:
          type: object
          description: UI hints to speed up review (optional).
        options:
          type: array
          description: For 'choice' type - options that apply patches to blocks when selected.
          items:
            type: object
            required: [id, label]
            properties:
              id:
                type: string
              label:
                type: string
              patches:
                type: array
                description: Patches that update blocks_working.
                items:
                  $ref: "#/components/schemas/BlockPatch"
        deny_reasons:
          type: array
          description: Quick deny reason suggestions shown as buttons in the deny modal.
          items:
            type: object
            required: [id, label]
            properties:
              id:
                type: string
                description: Machine-readable identifier for the reason.
              label:
                type: string
                description: Human-readable label shown on button.
          example:
            - id: "incorrect_info"
              label: "Incorrect information"
            - id: "wrong_recipient"
              label: "Wrong recipient"
            - id: "needs_revision"
              label: "Needs revision"
        external_url:
          type: string
          format: uri
          description: For 'external' type - URL to open for external review. User must confirm they completed external action.
        require_confirmation:
          type: boolean
          default: true
          description: If true (default), show confirmation modal before approve/deny.
    BlockPatch:
      type: object
      required: [block_id, op]
      properties:
        block_id:
          type: string
        op:
          type: string
          enum: [replace]
        value:
          oneOf:
            - type: string
            - type: object
    TextDiff:
      type: object
      required: [block_id, unified_diff]
      properties:
        block_id:
          type: string
        unified_diff:
          type: string
          description: Unified diff for markdown/plaintext blocks.
    JsonPatch:
      type: object
      required: [block_id, patch]
      properties:
        block_id:
          type: string
        patch:
          type: array
          description: RFC 6902 JSON Patch operations.
          items:
            type: object
            required: [op, path]
            properties:
              op:
                type: string
                enum: [add, remove, replace, move, copy, test]
              path:
                type: string
              from:
                type: string
              value: {}
    DecisionDiff:
      type: object
      properties:
        text_diffs:
          type: array
          items:
            $ref: "#/components/schemas/TextDiff"
        json_patches:
          type: array
          items:
            $ref: "#/components/schemas/JsonPatch"
    ReviewDecision:
      type: object
      required: [type, decided_at]
      properties:
        type:
          $ref: "#/components/schemas/DecisionType"
        reason:
          type: string
          description: Required when DENY; optional when APPROVE.
        decided_at:
          $ref: "#/components/schemas/Timestamp"
        decided_by:
          type: string
          description: Reviewer identifier (string for MVP).

    # -------- Source / Integration --------
    Source:
      type: object
      required: [id, name, created_at, delivery]
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        name:
          type: string
          description: Human-readable name for upstream integration/source.
        description:
          type: string
        created_at:
          $ref: "#/components/schemas/Timestamp"
        delivery:
          $ref: "#/components/schemas/SourceDeliveryConfig"
    SourceDeliveryConfig:
      type: object
      required: [mode, webhook]
      properties:
        mode:
          type: string
          enum: [WEBHOOK_ONLY, PULL_ONLY, WEBHOOK_AND_PULL]
        webhook:
          type: object
          required: [enabled]
          properties:
            enabled:
              type: boolean
            url:
              type: string
              description: Webhook URL for downstream executor (per source).
            secret:
              type: string
              description: Webhook signing secret (store hashed/secured).
            timeout_ms:
              type: integer
              default: 5000
            max_attempts:
              type: integer
              default: 10
            retry_backoff_seconds:
              type: integer
              default: 30

    # -------- Review Task --------
    ReviewTask:
      type: object
      required: [id, source_id, source_name, status, priority, title, service, action, risk_level, blocks_original, blocks_working, created_at]
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        source_id:
          $ref: "#/components/schemas/UUID"
        source_name:
          type: string
          description: Human-readable name of the source configuration (denormalized for display).
        status:
          $ref: "#/components/schemas/ReviewTaskStatus"
        priority:
          $ref: "#/components/schemas/Priority"
        title:
          type: string
          description: Task title/summary for display (max 200 chars).
        preview:
          type: string
          description: Preview of main content for card display (max 100 chars).
        service:
          $ref: "#/components/schemas/ServiceIdentifier"
          description: The target service where the action will execute (e.g., Gmail, Twitter).
        source_service:
          $ref: "#/components/schemas/ServiceIdentifier"
          description: For cross-service flows, the originating service (e.g., Notion notes → Gmail email).
        action:
          $ref: "#/components/schemas/ActionIdentifier"
          description: The action that will be performed if approved.
        risk_level:
          $ref: "#/components/schemas/RiskLevel"
          description: Risk level for UI friction scaling.
        risk_warning:
          type: string
          description: Warning message for high/critical risk tasks (e.g., "Irreversible financial transaction").
        interaction_schema:
          $ref: "#/components/schemas/InteractionSchema"
        blocks_original:
          type: array
          items:
            $ref: "#/components/schemas/ArtifactBlock"
        blocks_working:
          type: array
          items:
            $ref: "#/components/schemas/ArtifactBlock"
        blocks_final:
          type: array
          items:
            $ref: "#/components/schemas/ArtifactBlock"
        diff:
          $ref: "#/components/schemas/DecisionDiff"
        decision:
          $ref: "#/components/schemas/ReviewDecision"
        dispatch:
          $ref: "#/components/schemas/DispatchInfo"
        execution_intent:
          $ref: "#/components/schemas/ExecutionIntent"
          description: Extracted from execution_intent block for UI display.
        metadata:
          type: object
          description: Arbitrary metadata (trace_id, external ids, tags).
        created_at:
          $ref: "#/components/schemas/Timestamp"
        updated_at:
          $ref: "#/components/schemas/Timestamp"
        archived_at:
          $ref: "#/components/schemas/Timestamp"

    # -------- Review Task List Item (lightweight for queue) --------
    ReviewTaskListItem:
      type: object
      description: Lightweight task representation for queue listing (excludes full block content).
      required: [id, source_id, source_name, status, priority, title, service, action, risk_level, created_at]
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        source_id:
          $ref: "#/components/schemas/UUID"
        source_name:
          type: string
          description: Human-readable name of the source configuration.
        status:
          $ref: "#/components/schemas/ReviewTaskStatus"
        priority:
          $ref: "#/components/schemas/Priority"
        title:
          type: string
          description: Task title for card display (max 200 chars).
        preview:
          type: string
          description: Preview of main content for card display (max 100 chars).
        service:
          $ref: "#/components/schemas/ServiceIdentifier"
          description: The target service where the action will execute.
        source_service:
          $ref: "#/components/schemas/ServiceIdentifier"
          description: For cross-service flows, the originating service.
        action:
          $ref: "#/components/schemas/ActionIdentifier"
          description: The action that will be performed if approved.
        risk_level:
          $ref: "#/components/schemas/RiskLevel"
          description: Risk level for visual indication (card border color).
        risk_warning:
          type: string
          description: Warning message for high/critical risk tasks.
        block_count:
          type: integer
          description: Number of artifact blocks in the task.
        has_changes:
          type: boolean
          description: True if blocks_working differs from blocks_original.
        created_at:
          $ref: "#/components/schemas/Timestamp"
        updated_at:
          $ref: "#/components/schemas/Timestamp"

    # -------- Execution Intent (for UI display) --------
    ExecutionIntent:
      type: object
      description: Describes what will happen if the task is approved. Extracted from execution_intent artifact block.
      properties:
        kind:
          type: string
          enum: [command_template, mcp_tool_call, http_request, custom]
          description: Type of execution action.
        template_id:
          type: string
          description: Reference to executor-owned command/template.
        display:
          type: object
          description: Human-readable display properties for UI.
          properties:
            action_type:
              type: string
              description: Human-readable action type (e.g., "Send Email", "Update Config").
            target:
              type: string
              description: Human-readable target description (e.g., "sam@vendor.com").
            warning:
              type: string
              description: Optional warning message (e.g., "External recipient", "Production environment").
            icon:
              type: string
              description: Optional icon identifier for UI.
    DispatchInfo:
      type: object
      properties:
        last_delivery_status:
          type: string
          enum: [NOT_SENT, IN_PROGRESS, DELIVERED, FAILED]
        last_attempt_at:
          $ref: "#/components/schemas/Timestamp"
        attempt_count:
          type: integer
        # event_id is what downstream dedupes on (at-least-once delivery).
        decision_event_id:
          $ref: "#/components/schemas/UUID"

    # -------- Decision Event (Webhook + Pull) --------
...
    CreateSourceRequest:
      type: object
      required: [name, delivery]
      properties:
        name:
          type: string
        description:
          type: string
        delivery:
          $ref: "#/components/schemas/SourceDeliveryConfig"
    CreateReviewTaskRequest:
      type: object
      required: [source_id, title, service, action, risk_level, blocks]
      properties:
        source_id:
          $ref: "#/components/schemas/UUID"
        title:
          type: string
          description: Task title for display (max 200 chars).
        service:
          $ref: "#/components/schemas/ServiceIdentifier"
          description: The target service where the action will execute.
        source_service:
          $ref: "#/components/schemas/ServiceIdentifier"
          description: For cross-service flows, the originating service.
        action:
          $ref: "#/components/schemas/ActionIdentifier"
          description: The action that will be performed if approved.
        risk_level:
          $ref: "#/components/schemas/RiskLevel"
          description: Risk level for UI friction scaling. Agent should set appropriately based on action type and audience.
        risk_warning:
          type: string
          description: Warning message for high/critical risk tasks (e.g., "Irreversible financial transaction").
        priority:
          $ref: "#/components/schemas/Priority"
        interaction_schema:
          $ref: "#/components/schemas/InteractionSchema"
        blocks:
          type: array
          items:
            $ref: "#/components/schemas/ArtifactBlock"
        metadata:
          type: object
    PatchBlocksRequest:
      type: object
      required: [blocks_working]
      properties:
        blocks_working:
          type: array
          items:
            $ref: "#/components/schemas/ArtifactBlock"
    SubmitDecisionRequest:
      type: object
      required: [decision]
      properties:
        decision:
          $ref: "#/components/schemas/DecisionType"
        reason:
          type: string
...
paths:
...
  /sources:
    post:
      summary: Create a source (per-integration configuration)
      operationId: createSource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSourceRequest"
      responses:
        "201":
          description: Source created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Source"

    get:
      summary: List sources
      operationId: listSources
      responses:
        "200":
          description: Sources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Source"
  /sources/{source_id}:
    get:
      summary: Get a source
      operationId: getSource
      parameters:
        - in: path
          name: source_id
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: Source
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Source"

    patch:
      summary: Update source configuration
      operationId: patchSource
      parameters:
        - in: path
          name: source_id
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSourceRequest"
      responses:
        "200":
          description: Source updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Source"

  # ---------------- Review Tasks ----------------
  /review-tasks:
    post:
      summary: Enqueue a review task (agent-proposed action awaiting human gate)
      operationId: createReviewTask
      parameters:
        - in: header
          name: Idempotency-Key
          required: false
          schema:
            type: string
          description: Idempotency key for safely retrying task creation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewTaskRequest"
      responses:
        "201":
          description: Review task created
          headers:
            X-RateLimit-Limit:
              $ref: "#/components/headers/RateLimitLimit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/RateLimitRemaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/RateLimitReset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewTask"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"

    get:
      summary: List review tasks
      operationId: listReviewTasks
      parameters:
        - in: query
          name: source_id
          schema:
            $ref: "#/components/schemas/UUID"
          description: Filter by source configuration ID.
        - in: query
          name: status
          schema:
            $ref: "#/components/schemas/ReviewTaskStatus"
          description: Filter by task status.
        - in: query
          name: risk_level
          schema:
            $ref: "#/components/schemas/RiskLevel"
          description: Filter by risk level.
        - in: query
          name: service_id
          schema:
            type: string
          description: Filter by target service ID (e.g., "gmail", "twitter").
        - in: query
          name: action_type
          schema:
            type: string
            enum: [send, reply, post, create, update, delete, archive, schedule, transfer, notify]
          description: Filter by action type.
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: cursor
          schema:
            type: string
      responses:
        "200":
          description: Tasks page
          content:
            application/json:
              schema:
                type: object
                required: [items, total_count]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/ReviewTaskListItem"
                    description: Lightweight task items for queue display.
                  total_count:
                    type: integer
                    description: Total number of tasks matching the filter (for pagination display).
                  next_cursor:
                    type: string
                    description: Cursor for fetching the next page. Null if no more pages.
  /review-tasks/{task_id}:
    get:
      summary: Get a review task
      operationId: getReviewTask
      parameters:
        - in: path
          name: task_id
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: Review task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewTask"
  /review-tasks/{task_id}/blocks:
    patch:
      summary: Update blocks_working (always editable)
      operationId: patchReviewTaskBlocks
      parameters:
        - in: path
          name: task_id
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
        - in: header
          name: If-Match
          required: false
          schema:
            type: string
          description: Optional ETag for optimistic concurrency.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchBlocksRequest"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewTask"
  /review-tasks/{task_id}/decision:
    post:
      summary: Submit a decision (gatekeeper output triggers downstream execution on APPROVE)
      operationId: submitReviewTaskDecision
      parameters:
        - in: path
          name: task_id
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitDecisionRequest"
      responses:
        "200":
          description: Decision recorded; dispatch will occur per source delivery config.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewTask"

  # ---------------- Pull API (per source) ----------------
...